Rename to excalidraw-viewer.ts
