+++ web_src/js/features/file-view.ts
@@ -1,7 +1,7 @@
 import type {FileRenderPlugin} from '../render/plugin.ts';
 import {newRenderPlugin3DViewer} from '../render/plugins/3d-viewer.ts';
 import {newRenderPluginPdfViewer} from '../render/plugins/pdf-viewer.ts';
-import {newRenderPluginExcalidrawViewer} from '../render/plugins/excalidraw-view.ts';
+import {newRenderPluginExcalidrawViewer} from '../render/plugins/excalidraw-viewer.ts';
 import {registerGlobalInitFunc} from '../modules/observer.ts';
 import {createElementFromHTML, showElem, toggleElemClass} from '../utils/dom.ts';
 import {html} from '../utils/html.ts';
